Add History Tree
Added the ability to walk the git history from the plugin.
It will allow you to view old versions of the file and diff revisions
against the current working versions (and each other).

There is also, a simple fix for the branch window not selecting the
correct version of the branch for display.